Spliting componenent & useing state pass function & Clean code 40% & New Files structure
