+++ src/index.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import App2 from './FunctionalApp'
+// import App from './App';
+import App from './components/Apps/FunctionalApp'
 import reportWebVitals from './reportWebVitals';
 ReactDOM.render(
   <React.StrictMode>
-
-    
-    <App2  />
+   <App  />
   </React.StrictMode>,
   document.getElementById('root')
 );
